Split Aura and Timestamp inherents out of paint
