+++ client/consensus/pow/Cargo.toml
@@ -11,10 +11,10 @@ primitives = { package = "substrate-primitives", path = "../../../primitives/cor
 sr-primitives = { path = "../../../primitives/sr-primitives" }
 client-api = { package = "substrate-client-api", path = "../../api" }
 block-builder-api = { package = "substrate-block-builder-runtime-api", path = "../../../primitives/block-builder/runtime-api" }
-paint-timestamp = { path = "../../../paint/timestamp" }
 inherents = { package = "substrate-inherents", path = "../../../primitives/inherents" }
 pow-primitives = { package = "substrate-consensus-pow-primitives", path = "../../../primitives/consensus/pow" }
 consensus-common = { package = "substrate-consensus-common", path = "../../../primitives/consensus/common" }
 log = "0.4.8"
 futures-preview = { version = "0.3.0-alpha.19", features = ["compat"] }
+sp-timestamp = {  path = "../../../primitives/timestamp" }
 derive_more = "0.15.0"
