+++ paint/aura/src/mock.rs
@@ -62,7 +62,7 @@ impl system::Trait for Test {
 	type Version = ();
 }
 
-impl timestamp::Trait for Test {
+impl paint_timestamp::Trait for Test {
 	type Moment = u64;
 	type OnTimestampSet = Aura;
 	type MinimumPeriod = MinimumPeriod;
