+++ paint/timestamp/Cargo.toml
@@ -12,6 +12,7 @@ sr-primitives = { path = "../../primitives/sr-primitives", default-features = fa
 inherents = { package = "substrate-inherents", path = "../../primitives/inherents", default-features = false }
 support = { package = "paint-support", path = "../support", default-features = false }
 system = { package = "paint-system", path = "../system", default-features = false }
+sp-timestamp = { path = "../../primitives/timestamp", default-features = false }
 impl-trait-for-tuples = "0.1.3"
 
 [dev-dependencies]
@@ -28,4 +29,5 @@ std = [
 	"support/std",
 	"serde",
 	"system/std",
+	"sp-timestamp/std"
 ]
