+++ primitives/consensus/aura/Cargo.toml
@@ -11,6 +11,8 @@ codec = { package = "parity-scale-codec", version = "1.0.0", default-features = 
 rstd = { package = "sr-std", path = "../../sr-std", default-features = false }
 sr-api = { path = "../../sr-api", default-features = false }
 sr-primitives = { path = "../../sr-primitives", default-features = false }
+inherents = { package = "substrate-inherents", path = "../../inherents", default-features = false }
+sp-timestamp = {  path = "../../timestamp", default-features = false }
 
 [features]
 default = ["std"]
@@ -20,4 +22,6 @@ std = [
 	"rstd/std",
 	"sr-api/std",
 	"sr-primitives/std",
+	"inherents/std",
+	"sp-timestamp/std",
 ]
